serve/requirements.txt
fastapi
aiortc
av
numpy
python-dotenv
requests
torch
torchaudio
silero-vad
